Rename `carbon.ns` => `carbon.constants`
