+++ lua/carbon/buffer.lua
@@ -1,8 +1,8 @@
-local ns = require('carbon.ns')
 local util = require('carbon.util')
 local entry = require('carbon.entry')
 local watcher = require('carbon.watcher')
 local settings = require('carbon.settings')
+local constants = require('carbon.constants')
 local buffer = {}
 local internal = {}
 local open_cwd = vim.loop.cwd()
@@ -537,19 +537,19 @@ function buffer.create()
 end
 
 function buffer.clear_extmarks(...)
-  local extmarks = vim.api.nvim_buf_get_extmarks(data.handle, ns.hl, ...)
+  local extmarks = vim.api.nvim_buf_get_extmarks(data.handle, constants.hl, ...)
 
   for _, extmark in ipairs(extmarks) do
-    vim.api.nvim_buf_del_extmark(data.handle, ns.hl, extmark[1])
+    vim.api.nvim_buf_del_extmark(data.handle, constants.hl, extmark[1])
   end
 end
 
 function buffer.clear_namespace(...)
-  vim.api.nvim_buf_clear_namespace(data.handle, ns.hl, ...)
+  vim.api.nvim_buf_clear_namespace(data.handle, constants.hl, ...)
 end
 
 function buffer.add_highlight(...)
-  vim.api.nvim_buf_add_highlight(data.handle, ns.hl, ...)
+  vim.api.nvim_buf_add_highlight(data.handle, constants.hl, ...)
 end
 
 function buffer.set_lines(start_lnum, end_lnum, lines)
@@ -638,10 +638,10 @@ function internal.create_insert_move(ctx)
 end
 
 function buffer.focus_flash(duration, group, start, finish)
-  vim.highlight.range(data.handle, ns.hl_tmp, group, start, finish, {})
+  vim.highlight.range(data.handle, constants.hl_tmp, group, start, finish, {})
   vim.defer_fn(function()
     if buffer.is_loaded() then
-      vim.api.nvim_buf_clear_namespace(data.handle, ns.hl_tmp, 0, -1)
+      vim.api.nvim_buf_clear_namespace(data.handle, constants.hl_tmp, 0, -1)
     end
   end, duration)
 end
